refactor: move notes list to shared
